Move client creation functions into a separate file
